[tune] move `on_no_available_trials` related logic to a subclass under runner.
